+++ src/lib/environment/mod_test.rs
@@ -1015,12 +1015,12 @@ fn setup_env_empty() {
         plugins: None,
     };
 
-    setup_env(&cli_args, &config, "setup_env_empty1", None);
+    setup_env(&cli_args, &config, "setup_env_empty1", None, &mut vec![]);
 
     let mut value = envmnt::get_or_panic("CARGO_MAKE_TASK");
     assert_eq!(value, "setup_env_empty1");
 
-    setup_env(&cli_args, &config, "setup_env_empty2", None);
+    setup_env(&cli_args, &config, "setup_env_empty2", None, &mut vec![]);
 
     let delay = time::Duration::from_millis(10);
     thread::sleep(delay);
@@ -1046,7 +1046,7 @@ fn setup_env_skip_git() {
         plugins: None,
     };
 
-    let env_info = setup_env(&cli_args, &config, "setup_env_empty1", None);
+    let env_info = setup_env(&cli_args, &config, "setup_env_empty1", None, &mut vec![]);
     assert!(env_info.git_info.user_name.is_none());
 }
 
@@ -1067,7 +1067,7 @@ fn setup_env_skip_rust() {
         plugins: None,
     };
 
-    let env_info = setup_env(&cli_args, &config, "setup_env_empty1", None);
+    let env_info = setup_env(&cli_args, &config, "setup_env_empty1", None, &mut vec![]);
     assert!(env_info.rust_info.channel.is_none());
 }
 
@@ -1088,7 +1088,7 @@ fn setup_env_skip_crate() {
         plugins: None,
     };
 
-    let env_info = setup_env(&cli_args, &config, "setup_env_empty1", None);
+    let env_info = setup_env(&cli_args, &config, "setup_env_empty1", None, &mut vec![]);
     assert!(env_info.crate_info.dependencies.is_none());
 }
 
@@ -1138,7 +1138,7 @@ fn setup_env_cli_arguments() {
 
     envmnt::set("CARGO_MAKE_TASK_ARGS", "EMPTY");
 
-    setup_env(&cli_args, &config, "setup_env_empty1", None);
+    setup_env(&cli_args, &config, "setup_env_empty1", None, &mut vec![]);
 
     let value = envmnt::get_or_panic("CARGO_MAKE_TASK_ARGS");
     assert_eq!(value, "arg1;arg2");
@@ -1169,7 +1169,7 @@ fn setup_env_values() {
     assert_eq!(envmnt::get_or("MY_ENV_KEY", "NONE"), "NONE".to_string());
     assert_eq!(envmnt::get_or("MY_ENV_KEY2", "NONE"), "NONE".to_string());
 
-    setup_env(&cli_args, &config, "set_env_values", None);
+    setup_env(&cli_args, &config, "set_env_values", None, &mut vec![]);
 
     assert_eq!(envmnt::get_or_panic("MY_ENV_KEY"), "MY_ENV_VALUE");
     assert_eq!(envmnt::get_or_panic("MY_ENV_KEY2"), "MY_ENV_VALUE2");
@@ -1211,7 +1211,7 @@ fn setup_env_script() {
         "NONE".to_string()
     );
 
-    setup_env(&cli_args, &config, "set_env_values", None);
+    setup_env(&cli_args, &config, "set_env_values", None, &mut vec![]);
 
     assert_eq!(envmnt::get_or_panic("MY_ENV_SCRIPT_KEY"), "MY_ENV_VALUE");
     assert_eq!(envmnt::get_or_panic("MY_ENV_SCRIPT_KEY2"), "script1");